fixed upload image

+++ src/providers/diskStorage.js
@@ -13,15 +13,17 @@ class DiskStorage {
   }
 
   async delete(file) {
-    const filePath = path.resolve(uploadConfig.UPLOADS_FOLDER, file)
-
-    try {
-      await fs.promises.stat(filePath)
-    } catch {
-      return
+    if (file != null) {
+      const filePath = path.resolve(uploadConfig.UPLOADS_FOLDER, file)
+      
+      try {
+        await fs.promises.stat(filePath)
+      } catch {
+        return
+      }
+      
+      await fs.promises.unlink(filePath)
     }
-
-    await fs.promises.unlink(filePath)
   }
 }
